requirements.txt
django
django-registration
django-bootstrap3
pillow
pytz
python-dateutil
mock
